Fix powershell crashes on 32bit systems

+++ Toolkit/AppDeployToolkit/AppDeployToolkitMain.cs
@@ -58,7 +58,7 @@ namespace PSADT
 		private static extern IntPtr SendMessageTimeout(IntPtr hWnd, uint Msg, IntPtr wParam, string lParam, uint fuFlags, uint uTimeout, IntPtr lpdwResult);
 
 		[DllImport("shell32.dll", CharSet = CharSet.Auto, SetLastError = false)]
-		private static extern void SHChangeNotify(long eventId, uint flags, IntPtr item1, IntPtr item2);
+		private static extern void SHChangeNotify(int eventId, uint flags, IntPtr item1, IntPtr item2);
 
 		public static void RefreshDesktopAndEnvironmentVariables()
 		{
